requirements.txt
rumps>=0.4.0
PyYAML>=6.0
py2app==0.28.8
setuptools==70.3.0
python-telegram-bot>=20.7
toml>=0.10.2
pytest>=7.0.0
pytest-asyncio>=0.21.0